raop: Cleanup pollfd when the RTSP state enters TEARDOWN
It seems that STATE_DISCONNECTED never happens in usual cases,
which left pollfd uncleaned even after the socket fds were
closed in the current version. This led to PulseAudio hang up when
starting playback after the previous connection is torn down.

This patch fixes the issue, thus fixes the hang up issue in such
a case.

Also, control and timing fds are registered in rtpoll, so they
should be closed after rtpoll is disabled. This is what the TCP
implementation does, so this patch also follows the same way.